fix: fix returning type

+++ src/main/java/io/github/hyperf0rm/runner/model/Result.java
@@ -23,7 +23,7 @@ public class Result {
         this.id = id;
     }
 
-    public float getDuration() {
+    public long getDuration() {
         return duration;
     }
